kr3/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
pyjwt
python-dotenv
slowapi
pydantic
